+++ lbrynet/daemon/Daemon.py
@@ -1,5 +1,6 @@
 import binascii
 import logging.handlers
+import math
 import mimetypes
 import os
 import base58
@@ -1067,12 +1068,24 @@ class Daemon(AuthJSONRPCServer):
         """
 
         # on startup, the wallet or network won't be available but we still need this call to work
-        has_wallet = self.session and self.session.wallet and self.session.wallet.network
-        local_height = self.session.wallet.network.get_local_height() if has_wallet else 0
-        remote_height = self.session.wallet.network.get_server_height() if has_wallet else 0
-        best_hash = (yield self.session.wallet.get_best_blockhash()) if has_wallet else None
-        wallet_is_encrypted = has_wallet and self.session.wallet.wallet and \
+        has_wallet = self.session and self.session.wallet
+        downloading_headers = self.session.wallet.is_downloading_headers() if has_wallet else False
+        has_wallet_with_network = has_wallet and self.session.wallet.network
+        local_height = self.session.wallet.network.get_local_height() if has_wallet_with_network else 0
+        remote_height = self.session.wallet.network.get_server_height() if has_wallet_with_network else 0
+        best_hash = (yield self.session.wallet.get_best_blockhash()) if has_wallet_with_network else None
+        wallet_is_encrypted = has_wallet_with_network and self.session.wallet.wallet and \
                               self.session.wallet.wallet.use_encryption
+        headers_progress_percent = self.session.wallet.get_headers_progress_percent() if has_wallet else 0
+        headers_download_progress = headers_progress_percent \
+            if downloading_headers or headers_progress_percent == 100 \
+            else math.ceil((local_height / float(remote_height)) * 100) if remote_height else 0
+
+        progress = [{
+            'name': 's3-download',
+            'label': 'Downloading headers from S3',
+            'percent': headers_download_progress
+        }] if downloading_headers else []
 
         response = {
             'lbry_id': base58.b58encode(self.node_id),
@@ -1097,7 +1110,8 @@ class Daemon(AuthJSONRPCServer):
                 'blocks': local_height,
                 'blocks_behind': remote_height - local_height,
                 'best_blockhash': best_hash,
-            }
+            },
+            'progress': progress
         }
         if session_status:
             blobs = yield self.session.blob_manager.get_all_verified_blobs()